splitting frontend view to its own file
